rename quantum_lazy mod to hseries
