+++ tket2-hseries/src/extension.rs
@@ -1,4 +1,4 @@
 //! This module defines the Hugr extensions used by tket2-hseries.
 pub mod futures;
-pub mod quantum_lazy;
+pub mod hseries;
 pub mod result;
