+++ tket2-hseries/src/extension/hseries.rs
